ALKEMY/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
DateTime==4.5
greenlet==1.1.2
idna==3.3
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
urllib3==1.26.11
zope.interface==5.4.0
unidecode
psycopg2==2.9.3